// slaughter-game/src/app/shared/model/Animation/drawObjectOptions.ts
import { Point } from "../point";

export interface DrawObjectOptions{
    image: HTMLImageElement;
    displayWidth: number; 
    displayHeight: number; 
    context: any;
    frameWidth: number;
    frameHeight: number;
    speedX: number;
    speedY: number;
    coord: Point;
    name?: string;
}